requirements.txt
Django==4.2.10
django-allauth==0.61.1
pymongo==4.6.1
django-environ==0.11.2
python-dotenv==1.0.1
pillow==10.2.0
django-tailwind==3.8.0
django-crispy-forms==2.1
crispy-tailwind==1.0.0
django-cleanup==8.0.0
mutagen==1.47.0